+++ pyveg/src/gee_interface.py
@@ -4,21 +4,12 @@ Functions to help interface with GEE, in particular to download images.
 
 import os
 import shutil
-import requests
 from datetime import datetime
-from zipfile import ZipFile, BadZipFile
 from geetools import cloud_mask
 import cv2 as cv
 
 import ee
 ee.Initialize()
-
-from .image_utils import (
-    convert_to_bw,
-    crop_image_npix,
-    save_image,
-    combine_tif
-)
 
 if os.name == "posix":
     TMPDIR = "/tmp/"
@@ -90,50 +81,6 @@ def add_NDVI(image, red_band, near_infrared_band):
         print ("Something went wrong in the NDVI variable construction")
         return image
 
-
-def download_and_unzip(url, output_tmpdir):
-    """
-    Given a URL from GEE, download it (will be a zipfile) to
-    a temporary directory, then extract archive to that same dir.
-    Then find the base filename of the resulting .tif files (there
-    should be one-file-per-band) and return that.
-    """
-
-    # GET the URL
-    r = requests.get(url)
-    if not r.status_code == 200:
-        raise RuntimeError(" HTTP Error getting download link {}".format(url))
-    # DO NOT remove output directory and recreate it
-    #shutil.rmtree(output_tmpdir, ignore_errors=True)
-    os.makedirs(output_tmpdir, exist_ok=True)
-    output_zipfile = os.path.join(output_tmpdir,"gee.zip")
-    with open(output_zipfile, "wb") as outfile:
-        outfile.write(r.content)
-    ## catch zipfile-related exceptions here, and if they arise,
-    ## write the name of the zipfile and the url to a logfile
-    try:
-        with ZipFile(output_zipfile, 'r') as zip_obj:
-            zip_obj.extractall(path=output_tmpdir)
-    except(BadZipFile):
-        with open(LOGFILE, "a") as logfile:
-            logfile.write("{}: {} {}\n".format(str(datetime.now()),
-                                               output_zipfile,
-                                               url))
-            return None
-    tif_files = [filename for filename in os.listdir(output_tmpdir) \
-                 if filename.endswith(".tif")]
-    if len(tif_files) == 0:
-        raise RuntimeError("No files extracted")
-
-    # get the filename before the "Bx" band identifier
-    tif_filebases = [tif_file.split(".")[0] for tif_file in tif_files]
-
-    # get the unique list
-    tif_filebases = set(tif_filebases)
-
-    # prepend the directory name to each of the filebases
-    return [os.path.join(output_tmpdir, tif_filebase) \
-            for tif_filebase in tif_filebases]
 
 
 def ee_prep_data(collection_dict,
